+++ HISTORY.rst
@@ -12,3 +12,4 @@ History
 ------------------
 
 * Allow users to change the parameters when downloading NUTS shapefiles from Eurostat's API in wordcloud_map().
+* Allow users to change the sharpness of the regional borders by channging the DPI value used when creating the masks.